refactor(mouse): share TerminalSurfaces + hit_candidates from mouse.rs
